backend/requirements.txt
fastapi==0.110.1
starlette==0.37.2
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
email-validator==2.3.0
python-multipart==0.0.22
openai==1.99.9
httpx==0.28.1
stripe==14.4.0
pdfplumber==0.11.9
